+++ src/Email.hs
@@ -1,0 +1,9 @@
+module Email
+  ( -- * Email address
+    EmailAddress
+  , validate
+  , toByteString
+  ) where
+
+import Text.Email.Parser (EmailAddress, toByteString)
+import Text.Email.Validate (validate)