+++ src/Email/Unsafe.hs
@@ -1,0 +1,5 @@
+module Email.Unsafe
+  ( unsafeEmailAddress
+  ) where
+
+import Text.Email.Validate (unsafeEmailAddress)